Adding wpcom-helper.php file
As on the WordPress.com VIP platform the recommended way for activating the plugin is via theme's functions.php, the `photonfill_init` function hooked to `plugins_loaded` does not get triggered, as it's too late.

The wpcom-helper.php file gets included right after the plugin gets loaded. It currently contians only the call to `photonfill_init()` for setting up the plugin.